requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
streamlit
python-dotenv
duckduckgo-search
faiss-cpu
wikipedia
arxiv
numexpr